fix(java-sdk): remove test artifact notes field and fix quantity type to Double
- Remove `notes` field from CreateQuoteRequest and CreateAndSendRequest
  that was accidentally left by forward-compatibility test (field doesn't
  exist in JS canonical contract)
- Change all `Integer quantity` fields to `Double quantity` across
  LineItem, AddLineItemRequest, UpdateLineItemRequest, BundleItem,
  BundleItemInput, AddBundleLineItemRequest — matches JS `number` type
  and prevents silent truncation of fractional quantities

+++ packages/java-sdk/src/main/java/com/turbodocx/models/quote/AddLineItemRequest.java
@@ -8,7 +8,7 @@ public class AddLineItemRequest {
     private String productName;
     private Double unitPrice;
     private String billingFrequency;
-    private Integer quantity;
+    private Double quantity;
     private Double discountPercent;
     private String categoryId;
     private String categoryName;
@@ -24,8 +24,8 @@ public class AddLineItemRequest {
     public void setUnitPrice(Double unitPrice) { this.unitPrice = unitPrice; }
     public String getBillingFrequency() { return billingFrequency; }
     public void setBillingFrequency(String billingFrequency) { this.billingFrequency = billingFrequency; }
-    public Integer getQuantity() { return quantity; }
-    public void setQuantity(Integer quantity) { this.quantity = quantity; }
+    public Double getQuantity() { return quantity; }
+    public void setQuantity(Double quantity) { this.quantity = quantity; }
     public Double getDiscountPercent() { return discountPercent; }
     public void setDiscountPercent(Double discountPercent) { this.discountPercent = discountPercent; }
     public String getCategoryId() { return categoryId; }
